final restructure is done
