+++ src/sycl/kernels/lora/device/sgemm_lora_a_fwd_runner.hpp
@@ -35,14 +35,12 @@
   This runner unpacks a tile option tag (defined in sgemm_lora_a_fwd_types.hpp)
   into the template parameters of the shared pointer-array grouped-GEMM core and
   drives it end to end. It builds the grouped-GEMM metadata + pointer arrays on
-  device (SYCL kernels, one thread per segment; no host round-trip) and then
-  calls launch_group_gemm_lora_fwd.
+  device and then drives the GroupGemmLoraFwd launcher.
 
   The heavy CUTLASS type assembly + lifecycle (can_implement / get_workspace_size
   / initialize / run) and the device-side metadata build both live in the shared,
-  reusable core group_gemm_lora_launcher.hpp -- this file only layers the A-fwd
-  runner on top of it, so the core stays intact for the other LoRA kernels
-  (e.g. B-fwd) to reuse. The tile axis lives in sgemm_lora_a_fwd_types.hpp.
+  reusable core group_gemm_lora_launcher.hpp The tile axis lives in
+  sgemm_lora_a_fwd_types.hpp.
 */
 
 #pragma once
@@ -52,16 +50,19 @@
 
 #include <sycl/sycl.hpp>
 
-#include "sycl/kernels/lora/device/group_gemm_lora_launcher.hpp"
+#include "cutlass/cutlass.h"
+#include "sycl/kernels/lora/common/group_gemm_types.hpp"
+#include "sycl/kernels/lora/common/grouped_gemm_meta.hpp"
 #include "sycl/kernels/lora/device/sgemm_lora_a_fwd_types.hpp"
 
 namespace sgemm_lora_a_fwd_impl {
 
 //----------------- runSgemmLoraAFwd (one-shot entry) ------------------------//
 // High-level entry called from the per-(dtype, tile) .cpp files generated by
-// SGEMMLoraAFwdXe20.cmake. Builds the pointer-array grouped-GEMM metadata and
-// dispatches to the shared grouped-GEMM core with the tile config from
-// TileOpt and (alpha=1, beta=0) for the pure D = A @ B^T of LoRA-A.
+// SGEMMLoraAFwdXe20.cmake. Builds the pointer-array grouped-GEMM metadata on
+// device, assembles the CUTLASS Arguments via args_from_options(), and drives
+// the shared GroupGemmLoraFwd launcher with the tile config from TileOpt and
+// (alpha=1, beta=0) for the pure D = A @ B^T of LoRA-A.
 template <typename T, typename TileOpt>
 inline void runSgemmLoraAFwd(
     const torch::Tensor& input_x,
@@ -74,6 +75,7 @@ inline void runSgemmLoraAFwd(
     const int num_segments,
     sycl::queue& queue) {
   namespace core = at::native::xpu;
+  using Types = typename TileOpt::template Types<T>;
 
   const int K = static_cast<int>(input_x.size(1));  // input_dim
   const int N = stack_num * max_rank;               // output columns
@@ -81,7 +83,7 @@ inline void runSgemmLoraAFwd(
   const auto device = input_x.device();
 
   // Metadata + pointer arrays are built on device by SYCL kernels (one thread
-  // per segment) See group_gemm_lora_launcher.hpp.
+  // per segment). See common/grouped_gemm_meta.hpp.
   auto meta =
       core::build_grouped_gemm_meta(seg_indptr_i32, weight_indices_i32, N, K, num_segments, elem_bytes, device, queue);
 
@@ -89,16 +91,11 @@ inline void runSgemmLoraAFwd(
   auto b_ptrs = core::make_device_ptrs(weights, meta.b_off, queue);
   auto d_ptrs = core::make_device_ptrs(output, meta.d_off, queue);
 
-  // The shared core uses the modern (non-legacy) grouped path
-  // (MainloopXeL1StagedGroup), which auto-selects the XMX DPAS MMA atom and all
-  // gmem load/store/prefetch copy atoms per dtype.
-  core::launch_group_gemm_lora_fwd<
-      T,
-      typename TileOpt::TileShape,
-      typename TileOpt::ThreadLayout,
-      typename TileOpt::LayoutB,
-      TileOpt::PipelineStages>(
-      queue,
+  // Build the CUTLASS Arguments (args_from_options) separately from launching.
+  // Types<T> is the fully-assembled GroupGemmTypes<> bundle for this (dtype, tile);
+  // the shared core uses the modern grouped path (MainloopXeL1StagedGroup), which
+  // auto-selects the XMX DPAS MMA atom and all gmem copy atoms per dtype.
+  auto args = core::args_from_options<Types>(
       meta.problem_sizes,
       a_ptrs,
       b_ptrs,
@@ -111,6 +108,22 @@ inline void runSgemmLoraAFwd(
       num_segments,
       /*alpha    =*/1.0f,
       /*beta     =*/0.0f);
+
+  typename Types::Gemm launcher;
+
+  cutlass::Status status = launcher.can_implement(args);
+  TORCH_CHECK(
+      status == cutlass::Status::kSuccess,
+      "runSgemmLoraAFwd: can_implement failed: status=",
+      cutlassGetStatusString(status));
+
+  const size_t workspace_size = Types::Gemm::get_workspace_size(args);
+  auto workspace_opts = torch::TensorOptions().dtype(torch::kUInt8).device(device);
+  auto workspace = torch::empty({static_cast<int64_t>(workspace_size)}, workspace_opts);
+
+  status = launcher.run(args, workspace.data_ptr(), queue);
+  TORCH_CHECK(
+      status == cutlass::Status::kSuccess, "runSgemmLoraAFwd: run failed: status=", cutlassGetStatusString(status));
 }
 
 }  // namespace sgemm_lora_a_fwd_impl
