+++ src/sycl/kernels/lora/device/sgemm_lora_a_fwd_types.hpp
@@ -35,11 +35,13 @@
   This is the single place where a tile variant is registered: each option tag
   (e.g. LoraAFwdTileLarge) bundles the CUTE tile shape, the subgroup thread
   layout, the B layout, and the pipeline-stages count that the pointer-array
-  grouped GEMM consumes. The runSgemmLoraAFwd<T, TileOpt>() one-shot entry that
-  consumes these tags lives in sgemm_lora_a_fwd_runner.hpp, and the shared,
-  reusable grouped-GEMM core (type assembly, lifecycle, device-side metadata
-  build) lives in group_gemm_lora_launcher.hpp -- kept intact so the other LoRA
-  kernels (e.g. B-fwd) can reuse it.
+  grouped GEMM consumes, and exposes a Types<T> alias that binds those knobs
+  into the shared GroupGemmTypes<> traits bundle (common/group_gemm_types.hpp).
+  The runSgemmLoraAFwd<T, TileOpt>() one-shot entry that consumes these tags
+  lives in sgemm_lora_a_fwd_runner.hpp, and the shared, reusable grouped-GEMM
+  core (lifecycle + device-side metadata build) lives in
+  group_gemm_lora_launcher.hpp -- kept intact so the other LoRA kernels (e.g.
+  B-fwd) can reuse it.
 
   Adding a new tile is a two-step change:
     1) Define a new option tag here.
@@ -52,6 +54,7 @@
 #include <cute/layout.hpp>
 
 #include "cutlass/layout/matrix.h"
+#include "sycl/kernels/lora/common/group_gemm_types.hpp"
 
 namespace sgemm_lora_a_fwd_impl {
 
@@ -60,6 +63,9 @@ namespace sgemm_lora_a_fwd_impl {
 // [num_loras, N, K] row-major, which is ColumnMajor when viewed as B in the
 // A @ B^T grouped GEMM contract used here (the auto-selected copy atom
 // free-transposes it).
+//
+// Each tag also exposes Types<T> -- the fully-assembled GroupGemmTypes<> bundle
+// for element type T -- which is what the launcher template is instantiated on.
 
 // Canonical BMG grouped-GEMM tile (upstream 04_bmg_grouped_gemm):
 //   TileShape       = 256 x 256 x 32
@@ -71,6 +77,9 @@ struct LoraAFwdTileLarge {
       cute::Layout<cute::Shape<cute::_8, cute::_4, cute::_1>, cute::Stride<cute::_4, cute::_1, cute::_0>>;
   using LayoutB = cutlass::layout::ColumnMajor;
   static constexpr int PipelineStages = 2;
+
+  template <typename T>
+  using Types = at::native::xpu::GroupGemmTypes<T, TileShape, ThreadLayout, LayoutB, PipelineStages>;
 };
 
 }  // namespace sgemm_lora_a_fwd_impl
